Modify InvalidExpressionException's namespace
