+++ Desktop-Learning-Assistant/TagFile/TagFileService.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using DesktopLearningAssistant.TagFile.Model;
-using DesktopLearningAssistant.TagFile.Context;
 using DesktopLearningAssistant.TagFile.Expression;
 using Microsoft.EntityFrameworkCore;
 using System.IO;
